Move panel dispatchers to v2
