+++ libs/designer/src/lib/core/actions/bjsworkflow/add.ts
@@ -12,8 +12,8 @@ import {
   updateErrorDetails,
   updateNodeSettings,
 } from '../../state/operation/operationMetadataSlice';
-import type { RelationshipIds } from '../../state/panel/panelInterfaces';
-import { changePanelNode, openPanel, setIsPanelLoading } from '../../state/panel/panelSlice';
+import type { RelationshipIds } from '../../state/panelV2/panelTypes';
+import { changePanelNode, openPanel, setIsPanelLoading } from '../../state/panelV2/panelSlice';
 import { addResultSchema } from '../../state/staticresultschema/staticresultsSlice';
 import type { NodeTokens, VariableDeclaration } from '../../state/tokens/tokensSlice';
 import { initializeTokensAndVariables } from '../../state/tokens/tokensSlice';
