+++ libs/designer/src/lib/ui/CanvasFinder.tsx
@@ -3,7 +3,7 @@ import { useState, useMemo, useEffect, useCallback } from 'react';
 import { useSelector, useDispatch } from 'react-redux';
 import { useInternalNode, useNodes, useReactFlow } from '@xyflow/react';
 
-import { useIsPanelCollapsed } from '../core/state/panel/panelSelectors';
+import { useIsPanelCollapsed } from '../core/state/panelV2/panelSelectors';
 import { useIsGraphEmpty } from '../core/state/workflow/workflowSelectors';
 import { clearFocusNode } from '../core/state/workflow/workflowSlice';
 import { DEFAULT_NODE_SIZE } from '../core/utils/graph';
